backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.0.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
redis>=5.0.0
websockets>=11.0.0
python-dateutil>=2.8.0
cryptography>=40.0.0
bcrypt>=4.0.0